+++ src/ActionGameCore/Assets/CharacterControls/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## [0.10.2] - DEVELOP
+## [0.11.0] - DEVELOP
+### Changed
+- Changed the module to inherit from CharacterModuleBase.
+- Changed the movement process to the Module class.
+### Added
+- Added a parameter to make the Pull-up module operable on the ground.
+### Fixed
+- Fixed a bug where it would hop on slopes.
+- Fixed friction calculation.
 
 ## [0.10.1] - 2024-08-08
 ### Changed